refactor: component separation (#3748)
* refactor: component separation


* refactor: changed state name


* refactor: changed component


---------
